merge changes from networknext/configstore

+++ server.ps1
@@ -1,5 +1,7 @@
 param(
-    [switch][bool]$generate
+    [switch][bool]$generate,
+    [switch][bool]$SkipDeps,
+    [switch][bool]$BuildOnly
 )
 
 $Args = ""
@@ -9,13 +11,19 @@ if ($generate) {
 
 Push-Location $PSScriptRoot\server
 try {
-    Write-Output "Fetch deps..."
-    go mod vendor
-    if ($LastExitCode -ne 0) {
-        exit $LastExitCode
+    if (!$SkipDeps) {
+        Write-Output "Fetch deps..."
+        go get -u github.com/golang/protobuf/protoc-gen-go
+        if ($LastExitCode -ne 0) {
+            exit $LastExitCode
+        }
+        go mod vendor
+        if ($LastExitCode -ne 0) {
+            exit $LastExitCode
+        }
+        rm -Recurse -Force vendor/github.com/golang/protobuf
+        git submodule update --init --recursive
     }
-    rm -Recurse -Force vendor/github.com/golang/protobuf
-    git submodule update --init --recursive
 
     Write-Output "Generate meta.go..."
     .\protoc.exe --go_out=plugins=grpc:meta .\meta.proto
@@ -30,6 +38,10 @@ try {
     Write-Output "Generate TypeScript gRPC client..."
     Push-Location .\typescript
     try {
+        yarn
+        if ($LastExitCode -ne 0) {
+            exit $LastExitCode
+        }
         if (!(Test-Path src\api)) {
             New-Item -ItemType Directory -Path src\api | Out-Null
         }
@@ -55,14 +67,16 @@ try {
         exit $LastExitCode
     }
 
-    Write-Output "Running & testing server..."
-    $env:CONFIGSTORE_GOOGLE_CLOUD_PROJECT_ID="configstore-test-001"
-    $env:CONFIGSTORE_GRPC_PORT="13389"
-    $env:CONFIGSTORE_HTTP_PORT="13390"
-    $env:CONFIGSTORE_SCHEMA_PATH="schema.json"
-    .\server.exe $Args
-    if ($LastExitCode -ne 0) {
-        exit $LastExitCode
+    if (!$BuildOnly) {
+        Write-Output "Running & testing server..."
+        $env:CONFIGSTORE_GOOGLE_CLOUD_PROJECT_ID="configstore-test-001"
+        $env:CONFIGSTORE_GRPC_PORT="13389"
+        $env:CONFIGSTORE_HTTP_PORT="13390"
+        $env:CONFIGSTORE_SCHEMA_PATH="schema.json"
+        .\server.exe $Args
+        if ($LastExitCode -ne 0) {
+            exit $LastExitCode
+        }
     }
 } finally {
     Pop-Location
